Update dependencies and images
This commit updates various dependencies to their latest versions and replaces some image files in the `assets/images/banner/` directory. The `cloud_firestore`, `firebase_auth`, `firebase_core`, and `firebase_storage` packages are updated to their newer versions. Additionally, the `connectivity_plus` package is updated to version 2.3.8. The `PromoSlider` widget is also updated to use a `PageView` instead of `CarouselSlider`. Lastly, the `SignUpController` is updated to include the `name` field in the `UserModel`.