"Updated dependencies and fixed some widgets in the app"
This commit includes updates to several dependencies, including the addition of the `auto_height_grid_view` package. It also includes changes to various widgets in the app, such as modifying the `ECGridLayout` widget to use `AutoHeightGridView`, and updating the `ECProductCardVertical` widget to use a fixed height for the image. Additionally, there are changes to the `HomeScreen`, `Store`, and `UpdateProfileController` widgets.